Protect the user from qfinishing patches with empty/unsaved descriptions to allow for the fact that it the description that ends up in the repository after a qfinish is that which existed when the patch was last refreshed rather than the latest change to the patch description.
This means that to offer the "edit" option when empty descriptions
are found we have to be a little more elaborate.  We also need to check
if the commited summary is the same as the patch description and, if
not, make it so.

This will involve popping to the patch to be qfinished, refreshing it
and pushing back to the old top.  Of course, if the patch description is
empty we offer the user the chance to edit it first.

Signed-off-by: Peter Williams <peter_ono@users.sourceforge.net>