Fix problem of confusing directory decorations and add unresolved status ('U') to the set of statuses displayed.
Main file tree directories will be decorated as:
  'U' (unresolved after merge) if it contains any unresolved files
  'M' (modified) if it contains tracked files that have been modified,
      added, removed or deleted
  '?' (not tracked) if it contains only files that are not being tracks
      or the other files it contains are being ignored
  'I' (ignored) if it only contains files that are being ignored, and
  otherwise will be undecorated

Directories labelled 'I' will not be displayed when "show hidden" is
False.  The '.hg' directory will not be displayed even when 'show hidden'
is True.

If there is demand for the ability to view the contents of '.hg' a
separate mechanism will be provided.

Signed-off-by: Peter Williams <peter_ono@users.sourceforge.net>