[Bug 2837274] Incorrect display of MQ patch files
If the --git option to qrefresh is used and there are copied or renamed
files in a patch then the patch is in the unapplied state its files will be
reported incorrectly.

This is because the code that extracts file data from the patch file is
ignorant of the git extensions to diff.

Signed-off-by: Peter Williams <pwil3058@bigpond.net.au>