gwsmhg.desktop: Cleanup of Name and Category
The old menu entry title did not match the style from other titles.
Most menu entry titles are less than 20 letters and doesn't use double-dash.

I think "gwsmhg Mercurial GUI" is short and to the point, contains the name so
it is recognizable for the experienced user, but also describes it briefly for
new users. But the lowercase g looks a bit odd in the menu...

Icon should not include extension.

Category "Application" is not permitted by
http://standards.freedesktop.org/menu-spec/menu-spec-1.0.html#category-registry

Category "X-Fedora" is deprecated by Fedora.

Author: Mads Kiilerich <mads@kiilerich.com>
Acked-by: Peter Williams <pwil3058@bigpond.net.au>